[scan] Make sure inputs into `fn` are not `device_data` IR nodes
Fixes https://github.com/pytorch/xla/issues/8742.

This is to workaround a limitation of `mark_sharding`, which replaces
the innards of the tensors it operates on. In other words, `mark_sharding`
is an in-place operation as opposed to a transform like found in JAX.

When `fn` contains a `mark_sharding` and the `mark_sharding` operates on one
of the carry or xs fake tensors, the original device data will be discarded
and a new one will be created in its place. As a result, `fn` will appear to
create a few empty tensors internally that are unrelated to the carry and xs
fake tensors, and the carry and xs will appear completely unused.

See https://github.com/pytorch/xla/issues/8742 for the bug. In short,
if an input into the layer to be scanned is a device data, and that layer
does a `mark_sharding` on said input, then the graph capturing in `scan`
will fail.

The workaround here is simple and cursed: multiply any `device_data` by 1.
This will make sure these tensor don't hold device data IR nodes and will
defeat the device data replacement of `mark_sharding`.

Fortunately, XLA simplifies away the multiplication (see [1]) so this should
become a no-op by the time it hits the TPU.

[1]: https://github.com/openxla/xla/blob/869f57d0082d7adbb9efc10cc18f51a562fc7bf3/xla/hlo/transforms/simplifiers/algebraic_simplifier.cc#L4755-L4770
